Create article for absolute value
I hope this can at least be a rough outline for the absolute value article. The markdown I used seems to be wrong somehow, but I can't figure out why. For example, if I delete all the text from the article the preview still shows text in it. So, it doesn't seem to be giving me an accurate preview as far as I can tell. Also, I couldn't figure out how to remove the green vertical bars from the left side of the text that's showing in my preview, and I wasn't sure whether to add a section on the LaTeX syntax of the absolute value function. Maybe someone can edit the article, or just let me know how to fix it?