Deploy 11 08 17 (#152)
* Add pre-commit hook to run lint and normaliseArticles (#102)

* Add article insert into statement

* Add Git stash article

* Add Git branch article

* Update index.md (#106)

* add info and example for articles (#107)

* add info about index

* add info about href

add information and exampe about he use of <a href>

* add concept and info of attributes and comments

definition and concept with an example at each article

* add concept and examples

the concepts and examples used are easy to understand and fulfill with
the requirements previously mentioned.

* Add SQL Insert Into Select Statement article (#117)

* Add Git clone article (#115)

* add information and example to the article (#116)

* add info about index

* add info about href

add information and exampe about he use of <a href>

* add concept and info of attributes and comments

definition and concept with an example at each article

* add concept and examples

the concepts and examples used are easy to understand and fulfill with
the requirements previously mentioned.

* add info and example

add information and example

* add concept and examples

concept with examples to improve the article

* hasOwnProperty article

* after normalize

* Add SQL or operator article (#128)

* Add SQL or operator article

* minor wording edit

* Add SQL not operator article (#127)

* final draft sql and operator (#126)

* final draft sql and operator

* Minor copy edits

* rebuild missing guide (#125)

* sql between final draft (#124)

* sql between final draft

* minor copy edits

* final for avg function (#123)

* final for avg function

* Minor copy edits

* Add SQL delete statement article (#120)

* Create article for absolute value (#121)

I hope this can at least be a rough outline for the absolute value article. The markdown I used seems to be wrong somehow, but I can't figure out why. For example, if I delete all the text from the article the preview still shows text in it. So, it doesn't seem to be giving me an accurate preview as far as I can tell. Also, I couldn't figure out how to remove the green vertical bars from the left side of the text that's showing in my preview, and I wasn't sure whether to add a section on the LaTeX syntax of the absolute value function. Maybe someone can edit the article, or just let me know how to fix it?

* Add SQL Update Statement article (#119)

* modified object-prototype-hasownproperty and removed meta command from README

* Update Adding and Subtracting Negative Numbers article (#133)

* Update Add Fractions with Unlike Denominators article (#132)

* Update index.md for all factors of a number article (#131)

* Removed stub from iframes article

* Remove object-prototype-__proto__ article

* [WIP]DevDocs-ify SideNav (#129)

* Drill down menu

* Update site title on navigation and add canonical link

* remove shouldComponentUpdate

* Fix SQL alter table article and normalise

* Refactor for Feedback

* final CREATE INDEX STATEMENT (#142)

* final CREATE INDEX STATEMENT

* Update index.md

* final draft COUNT AGGREGATE FUNCTION (#141)

* final draft COUNT AGGREGATE FUNCTION

* Minor copy edits

* Fix Gatsby SSR (HTML Build) (#150)