Fixed broken link and updated compile instruction (#311)
* Fixed broken link

The hyperlink to 'Hello World' program explanation in the Java documentation wasn't properly placed. Fixed it.

* Updated compile instruction for Hello World program

A novice programmer might not be familiar with the use of terminal / Command Prompt and may not know that you need to be in the directory the file is contained in while compiling and running the code.