Updating git checkout (#507)
* Add article for Git Checkout

Hi,  This is my first pull to this project.  If there is something that needs to be added or changed please let me know and I am more than happy to update it. 
Thanks,
Mary

* updated with feedback

Thank you so much for the feedback.  It has now been updated with all of the revisions you requested.  Let me know if there is anything else that needs to be changed or updated and I will be sure to fix it. 

Thanks,

Mary

* fixing small revision

Missed a small revision.

* Text and formatting edits

* Syntax highlighting