Index.me
Visual Basic for Applications is the programming language created by Microsoft for use within in their applications; Office, Word, Excel, and Access. It is an event-driven language and can be used to build user-defined-functions(UDFs), access windows API, automation(OLE), and dynamic-link-library's(DLLs). However VBA is not a standalone language and cannot be used outside of a Microsoft application. VBA has many uses, in the office space it is often used to automate tasks and generate reports. Since it is a easy language to grasp it is user-friendly and easy to implement into almost any office task.