Modified Title & Definition
'Jump Statements' should be the correct title for this page, as control statements broadly consist of selection (if, if-else); iteration (for, while, do-while); and jump statements (break, continue). Added the corresponding definition.

Need to update the folder name accordingly.

Also, the articles on 'break' and 'continue' statements need to be placed WITHIN this folder. They are currently at the same level as this directory.