Create an entry for Electron
I use Electron and thought it seemed like a good fit for the guide. I searched for "electron" in all the PRs and issues and didn't find anything, so I figured it would be safe to open this. This is my first commit, though, so let me know if I've done anything wrong. Thanks!