Create a folder called Bootstrap-popover
Bootstrap-popover is a small overlay of content that is used to display secondary information of any element when it is clicked by a user.
For an e.g. if a user is not logged in and trying to perform a task a for which he must be logged in, a bootstrap-popover can be used to show the message to the user. Hence it is a very useful component of bootstrap and should be included in the guide.