added the text " is the creation... towards making it better and easier"
Mobile application development is the creation of apps or programs that will run on a smartphone or tablet.  The two common platforms of development in the mobile app industry are IOS or Android,  IOS can be written in Objective-C or the new programming language of Swift.  Andoird is mainly developed in the programming language of Java.  Some major tools developers use for development in this field is X-code for IOS development, and Android Studio for Android development.  These two are the major ones used in the industry due to their versatility and well built community that is using it and continously proposing ideas and solutions towards making it better and easier.