added the text "C++ was an incredible...edition of C++"
C++ was an incredible programming language designed by Bjourne Stroustrap.  It introduced a wide variety of materials to the programming world and programming languages.  As a result due to its highly sophisticatedness,  C++ set a precedent for future programming languages, and many were completely based upon C++.  Some of these languages include C#, Java, and Python.  Overall thanks to Bjourne Stroustrap the programming world was given an excellent gift with the edition of C++.