added information regarding ASP.NET
Added information regarding the .NET framework and how that works with C#. I want to add a topic called 'ASPNET' as a folder and write about the framework in more detail, then add a link back here to the ASPNET page. With the open sourcing of .NET code in 2008 and .NET applications becoming machine independent, it has become a relevant platform, especially in business use cases, so it should at least get some airtime here.