Added proof to quadratic formula
Other people have already written detailed explanations of the quadratic formula but they haven't been merged yet. I just wanted to add a proof of the quadratic formula to whatever has already been written because I think its an important part of the explanation and will help people understand it better. 

Apologies if this is the wrong place to do this, not sure how to go about adding to articles not merged yet I just want this to be added on to the end to the accepted pull request. 
If anything in my proof needs to be altered I will also leave the original file here 

[Quadratic Formula Proof.docx](https://github.com/freeCodeCamp/guides/files/1420006/Quadratic.Formula.Proof.docx)