Added
This guide will demonstrate "How to Create an HTML Button That Acts Like a Link".
Programmers and Web Developers use buttons for several different ocassions. 
Today you will create your very own social link button! 

## Step 1 Button Tag

No not a HashTag, but a button tag. This is the first step and it is the step that defines to the computer
what you want to see. Lets Start!

<button>
  
# BEST Practice

You should get in the habit of closing your tags first before writing code. So as to not run into unwanted errors or your code just not working at all. It may seem silly but when you have alot of code written down it can be easy to forget to close your tags or divs. Resulting in bad or not working code.

So lets close the button tag, like so...

<button></button>

## Step 2 Anchor Tag

You know the anchor Tag, <a> well if you don't its the tag that you can use to bring up a local (on your computer) or network (internet) file or page. For this demonstration you will need to put an anchor element within your button tag.
  
<button><a href="Put_Your_Facebook_link_here_"</a>Add Text Here Such As "Twitter"</button>

## Test Your Code

Before you test you test your new social link button. Look over your code for errors such as an unclosed tag or div and or misspelled words. Right now this may seem silly since its only a 1 line code but as you start learning and coding more and more you will find this more of a daunting task. Which is why its Best Practice to close your tags  and divs before you start coding. 

## You Did IT 

You did it, you created a social link button or better yet a button that acts as a link!

Your code should look like somewhat like this.

<button><a href="https://twitter.com/_Jameslemon">Twitter</a></button>

To get rid of that underline simply input the following: style="text-decoration: none;" between the href link and the close tag. Like this, 

<button><a href="https://twitter.com/_Jameslemon" style="text-decoration: none;">Twitter</a></button>