Revised Changes Made to Article
This guide will demonstrate "How to Create an HTML Button That Acts Like a Link".
Programmers and Web Developers use buttons for numerous different occasions. Some Buttons are for submit forms such as emails, newsletters, or polls and etc. While others are to open links or bring up pages.

Today you will create your very own social link button!

## Button Tag

This is the first step and it is the step that defines to the computer
what you want to see. Lets Start! Type the following as shown below.

<button>

## Best Practice

It is good to get in the habit of closing your tags first before writing code. To avoid any syntax errors and to insure your code runs smoothly. It may seem unimportant with the line of code demonstrated in this guide.

However there will be times when you will have a numerous lines. It is very simple to forget to close your <tag>s or <div>s resulting in bad or not working code.

So lets close the button tag, like so...

<button></button>

## Anchor Tag

An anchor tag is one that you can use to bring up a local (on your computer) or network (internet) file or page. For this demonstration you will need to put an anchor tag within your button tag.

<button><a href="Put_Your_Facebook_link_here_"</a>Add Text Here Such As "Twitter"</button>

## Test Your Code

Before you test your new social link button. Look over your code for errors such as an unclosed <tag> or <div> and or misspelled words.

## You Did IT

You did it, you created a social link button or better yet a button that acts as a link!

Your code should look like somewhat like this.

<button><a href="https://twitter.com/_Jameslemon">Twitter</a></button>

To get rid of that underline simply input the following: style="text-decoration: none;" between the href link and the close tag. Like this,

<button><a href="https://twitter.com/_Jameslemon" style="text-decoration: none;">Twitter</a></button>