Create a new folder in CSS for CSS preprocessors (#1597)
LESS and SASS are core to a lot of modern web development and key things for new developers to try and be familiar with. Short guides would be super helpful! But before diving into specific preprocessors, giving new learners a sense of what a CSS preprocessor is and what it does is really important.