Working In Tech: Fixing book links
The book links to Amazon were not wrapped in anchor tags so a visitor would have to copy and paste the link into a new tab or window. I wrapped the word Amazon so the user can click the word and be taken to the Amazon page for the book in a new window.