Overview write up of the basic editor interface
Brief overview of the essentials of the unity editor window. Feedback welcome.

Not sure if this is the correct way to do a PR for a stub for the file folder that should be above the article I'm trying to add, and then the article I'm trying to add, but I did them in separate PRs so potentially the stub could be accepted and this rejected and to have separate feedback.