Updated Clojure/Conditionals with `cond` and forms
`cond` is a common conditional operator so I added it to the **Alternatives to If** section. 

I also added a section on special forms. Conditional expressions are a good place to introduce the concept of special forms. Their evaluation rules are intuitive for learners to grasp, but still clearly contrast with Clojure/Lisp's normal evaluation rules. 

I also fixed a small grammatical error.