Add progressive web app section and sub topics (#7159)
* Adding a section for progressive web apps

Adding the new section and the start of the article

* Delete index.md

* add new section for progressive web apps

Adding the new section covering progressive web apps and the beginning of the article for progressive web apps and sections for service workers and app shell

* changed stub title and section

* changed the stub title and first section

* Updated stub article url

* Updated stub url

* Updated two sections and added citation

I updated the first paragraph to include a longer description of what I understand PWAs to be. The second section is directly quoted from an engineer, Alex Russel who defined and helped standardize what PWAs are and how they should function. I am citing from his original blog post from June 15, 2015.

* Added citation to blockquote for attribution clarification

I added the list of PWA attributes and single sentence quote to the same blockquote to clarify they are from the first source cited in the sources section.

* Update index.md

* Update index.md

* Wording and additional information