Fixed issue where photo was not displaying
Currently I have the photo being hosted offsite at my Cloudinary, since for some reason the image path was not properly entered in the article, and thus would redirect to to the github codebase page, and would also display as a link rather than an image. 
I'm not entirely sure why the first two div lines are necessary, but upon removing them, the raw text is actually displayed, but when in place they are hidden when viewing the page. I will look more into this. Any help with hosting the image onsite would be greatly appreciated.