Started the article for "Recursive Formulas for Arithmetic Sequences" (#7840)
* Started the article for "Recursive Formulas for Arithmetic Sequences"

The bulk of the article is written. The following subsections should be added:
- Given a sequence, how do I construct the formula
- For more information, check out these links
    - Link to a khanacademy video
    - Link to the wikipedia for Arithmetic Sequences
    - See the complementing article on "Recursive Formulas for Geometric Sequences"
        - Make a stub for "Recursive Formulas for Geometric Sequences"

* Arithmetic Sequences: Finished article.

All changes mentioned in the previous commit have been added.
The sister article "Recursive Formulas for Geometric Sequences" still needs to be created and then this guide should be updated with that link.

* Arithmetic Sequences: Switched focus to Recursion

The two initial commits were too focused on what an arithmetic sequence was in general. This new version gives a quick introduction to arithmetic sequences, then goes into the recursive part.