Update index.md to address some typos and awkward phrasings (#5143)
I have proposed some edits to the index.md file that should enhance readability. Most are corrections of simple typographical errors, but some are ways to rephrase statements to increase clarity. I have additionally added a link to a StackOverflow page describing the relationship between Python and TensorFlow, which was slightly misrepresented in the prior draft.