Update index.md
The information was not so descriptive.Moreover the in general the three parts of a for loop statement is defined as initilization,condition and update. In the previous description the name of update expression was declared incorrectly as 'increment' which becomes a special case because the update operation can also be applied for decrementation (i.e : a-- operator).Moreover I also added an explanation for the code because an explanation is always needed whenever you are writing a guide.