Added text to line 7 (#9516)
* Added text to line 7

"The main feature that separates Adobe Illustrator from Photoshop is the use of vector graphics. Vector graphics construct images using mathematical statements allowing the images to cleanly scale up and down in size without image quality reduction. Adobe Photoshop uses raster graphics that focuses on a set pixel size canvas. If one was to scale a raster image up or down, the results would become pixelated, causing blurry or rough edges. It is for this reason that Adobe Illustrator is typically used for large and small designs ranging from software icons to billboards. "

* Update index.md