C++ syntax highlightning and small naming fix
I would like to suggest some small correction in code block in order to highlight c++ syntax.
Also it will be correct to say `Printing string` instead of `Printing character` because 'character' corresponds to a single symbol, not a sequence of symbols like in the code example. So I changed method input parameter type to ` string` and output text.