Updated solution for new certification (#8645)
There have been changes in respect of the previous course.
For example, now the changes to the code start later, and the previously proposed solutions work in the new course but require to edit the code out of the boundaries set by the lesson. Also, this solution uses the previously learned indexOf() instead of splice().