Adding instructions to fix .gitignore mistakes (#21334)
* Adding instructions to clean up .gitignore errors

Inexperienced devs often run into these situations that can be frustrating and a little scary to fix. The information is easy enough to find, and I hate to overcomplicate these instructions, but since these are common enough mistakes for the newly initiated, it may be good to have a reference here.

* Adding instruction to fix .gitignore mistakes

A common mistake for the inexperienced is to commit files and directories without updating the .gitignore file. This information is easily found if you dig for it, but common enough I thought it might be useful to add here.